June 27, 2016
##Updates##
images located in /images moved to new folder /images/site: This had to
be done in order to facilitate differentiation and manipulation of these
images on the webserver due to the way ExpressionEngine is configured at
this time. All other files from 'Static Site' uploaded to their
respective folders and the site mapping notes were added to the
README.md file. 'Static SIte' changed to 'Static Site' because it made
my eye twitch a little.